Add type checks for RtTypedValue access in debug builds.
