+++ source/MaterialXRuntime/RtType.cpp
@@ -1,0 +1,42 @@
+//
+// TM & (c) 2021 Lucasfilm Entertainment Company Ltd. and Lucasfilm Ltd.
+// All rights reserved.  See LICENSE.txt for license.
+//
+
+#include <MaterialXRuntime/RtType.h>
+
+namespace MaterialX
+{
+
+const RtIdentifier RtType::BOOLEAN("boolean");
+const RtIdentifier RtType::INTEGER("integer");
+const RtIdentifier RtType::FLOAT("float");
+const RtIdentifier RtType::VECTOR2("vector2");
+const RtIdentifier RtType::VECTOR3("vector3");
+const RtIdentifier RtType::VECTOR4("vector4");
+const RtIdentifier RtType::COLOR3("color3");
+const RtIdentifier RtType::COLOR4("color4");
+const RtIdentifier RtType::MATRIX33("matrix33");
+const RtIdentifier RtType::MATRIX44("matrix44");
+const RtIdentifier RtType::IDENTIFIER("identifier");
+const RtIdentifier RtType::STRING("string");
+const RtIdentifier RtType::FILENAME("filename");
+const RtIdentifier RtType::INTEGERARRAY("integerarray");
+const RtIdentifier RtType::FLOATARRAY("floatarray");
+const RtIdentifier RtType::COLOR3ARRAY("color3array");
+const RtIdentifier RtType::COLOR4ARRAY("color4array");
+const RtIdentifier RtType::VECTOR2ARRAY("vector2array");
+const RtIdentifier RtType::VECTOR3ARRAY("vector3array");
+const RtIdentifier RtType::VECTOR4ARRAY("vector4array");
+const RtIdentifier RtType::STRINGARRAY("stringarray");
+const RtIdentifier RtType::BSDF("BSDF");
+const RtIdentifier RtType::EDF("EDF");
+const RtIdentifier RtType::VDF("VDF");
+const RtIdentifier RtType::SURFACESHADER("surfaceshader");
+const RtIdentifier RtType::VOLUMESHADER("volumeshader");
+const RtIdentifier RtType::DISPLACEMENTSHADER("displacementshader");
+const RtIdentifier RtType::LIGHTSHADER("lightshader");
+const RtIdentifier RtType::MATERIAL("material");
+const RtIdentifier RtType::AUTO("auto");
+
+}
